Allowed nested types to be used in functions in the same class.
Previously, types used inside classes were resolved before anything was added to
the class. This was fine as long as no types could be declared inside the
classes. However, since we now allow nested type declarations, we need to be
more careful to do things in the right order.

Also fixed a bug in the Private access modifier that accidentally examined one
node 'too deep' in the type tree when looking for types. This resulted in
private nested types not being visible from the surrounding type.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>