Fixed an issue in function overloading.
When a subclass overloaded a function in a parent class with multiple possible
options (eg. by takin an Object as parameter), the compiler would bail out since
it saw multiple possibilities and did not prioritize them. Now, it will consider
an exact match to be better than an inexact match, but inexact matches to be
ambiguous in order to not cause any surprising behaviour.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>