Improved robustness in function calls and inline functions.
Previously, it was possible to pass registers as parameters to function
calls. If the call in question was an inlined function or a thread call, this
could cause problems as these types of functions assumed that all parameters
were in memory (which was previously enforced by only accepting Array<Var>).

Now, the InlineParams interface has been changed to make it easy to handle the
case where not all parameters are located in memory by moving the responsibility
of loading parameters into registers to InlineParams. Since InlineParams is
aware that some parameters may reside in registers, it can graciously handle
those cases without having to place such logic inside all inlined functions.

The thread call logic has been updated in a similar fashion. In this case, all
parameters in registers are spilled since the implementation needs to call some
functions (eg. constructors) before parameters can be stored in their final
location.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>