Fixed a performance bug accidentally introduced earlier.
The error caused the cache inside VTableCall to not function properly, which in
turn degraded performance quite a bit. This shows that code generation is
fairly expensive at the moment.

The issue was fixed, and a hint to the GC was inserted in
CppLoader::loadFunctions(), which together reduced startup times by around 30-40
ms on a 32-bit Windows machine.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>