Resolved crash on GCC 8.1.0.
When compiling Storm with GCC 8.1.0, the test suite started crashing. This was
due to GCC ignoring the global -falign-functions=2 flag, at least for the static
template functions in CppTypes.cpp. Since this caused some functions to reside
at an odd address, the thread dispatcher erroneously attempted to use vtable
dispatch in those cases, which caused a crash. This is fixed by adding the
appropriate attribute to the CODECALL macro for GCC on X64.

According to the GCC documentation, it is possible that this does not happen
when using -O2 since the documentation states that -falign-functions is implied
there. On the other hand, since -falign-functions does not seem to work in this
particular case, I don't know about that.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>