Found a case where the Basic Storm assignment operator failed.
This could happen since automatic type casting was allowed for all parameters
during a function call, even the 'this'-parameter. This could, for example, lead
to an expression like this:

Foo? x; Foo y; x = y;

To compile properly, but failing to produce the desired outcome.

This was triggered by updating the signature of NameLookup::parent() to return
MAYBE(T) instead of just T.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>