Fixed a bug in the hash tables.
If a "put" operation caused a rehash, the hash tables would erroneously clear the location
dependency for the objects inside the table, most importantly, the newly added object was
forgotten. This caused stale hash values in the tables, which leads to elements not being found in
hash tables, and thereby other strange behaviours (such as non-functioning vtables).

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>