Improved robustness of the pattern syntax.
Previously, we assumed that the contents of a pattern block would be transformed
once, as if everything would be transformed by the built-in transform
functions. This is not always the case, since parts of the parse tree may be
transformed multiple times and out of order (eg. when dealing with lambda
functions). Now, the pattern system keeps track of the proper index in each
node, so that nodes may be transformed multiple times without issues.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>