Fixed a stack layout issue in the X64 backend.
This was due to an incorrect assumption in the implementation of the < operator for Size and
Offset, which caused the size calculation of stack frames to fail from time to time, resulting in
corruption of local variables.

Also: optimized the constructor code in Basic Storm by omitting a temporary variable used to run
destructors. When the type in question does not have a destructor, the variable can safely be
omitted, saving stack space and omitting two machine instructions.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>