Fixed an crash when calling EnginePtr functions with the X64 backend.
The code generating thunks for functions taking an EnginePtr would fail to take
into account the case where the function returns complex types. This would lead
to the EnginePtr value being injected as the hidden 0th parameter instead of the
first as expected, which in turn would cause strange crashes. One example of
such a function is 'core.io.buffer()'.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>