Moved the return type checks for overloads to Type.
This means we can spot mistakes in return type that would previously cause the
function overload to be completely ignored. Furthermore, this solves an issue
that allowed an overloaded function to return "void" even if the parent function
did not, which caused strange issues.

This change revealed a number of such errors in the implementation of Basic
Storm, which were fixed.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>