Fixed a use-after-free issue in the system.
Since the MPS uses two step finalization, storing pointers to object with
destructors in a WeakSet could result in gaining access to finalized objects
when iterating through the WeakSet. Objects that have been finalized are still
present since they have been kept alive by the finalization message and it is
possible that they are collected at a much later time.

This could cause the UI library to crash during shutdown when resource objects
were created and discarded in a rapid fashion.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>