Re-worked the picture block into a custom block with its own syntax.
Now it is easier to customize the behaviour of things in the picture block, as
they are just a declarative wrapper around an interactive scheme rather than a
dynamic layout system.

We also modified Basic Storm to look for member functions with the same name as
a constructor. Otherwise we would not find Node:north since we have the class
north.

Also fixed the Emacs plugin. Previously, when the compiler generated a
compilation error and did not respond to the autocomplete messages, we would
store a symbol in the lookup cache, which caused Emacs to attempt to call it as
a function.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>