Fixed re-parsing issue in the language server.
When adding statements in a large block, the language server would not find a
large enough portion to re-parse, which caused weird results. However, it did
not trigger a full re-parse as one would expect, which usually caused newly
entered statement to appear as comments. This is now fixed by scheduling a large
re-parse in these situations. This was most notable when creating large
presentations, or large syntax files.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>