Optimized the creation of GcType objects a tiny bit.
Avoid zeroing memory on C++11 and onwards by providing a default constructor
that does not initialize the "type" member of MpsType. This means we could
remove the mitigation for the same problem in "gcTypeSize", but since that
mitigation does not waste a great deal of memory, it is better to be on the safe
side if some compiler decides it is cheaper to initialize the entire MpsType to
something rather than touching just the first two pointers (unlikely, but it
will be a pain to debug).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>