Implemented lookup from address to blocks in BlockAlloc.
This is to provide the ability to notify blocks when they have been
modified. This ability is necessary on Linux. However, it is not strictly
necessary on Windows, but likely speeds up write detection there as well, since
we will be able to query for writes once for the entire address range rather
than for each block that we examine.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>