Using GetWriteWatch to get notified about writes to objects.
This is also used inside Block, to know if the summary is stale. This requires
care at the moment, as we only update the staleness of references when the arena
is entered, not during a scan (at least on Windows).

This mechanism is not implemented at all for Posix-based systems yet.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>