SMM supports allocating GcType descriptions.
These are handled almost like regular objects, i.e. they do not require special
garbage collection as is the case with the MPS. This is solved by the special
case in the formatting introduced previously that preserves most of the data in
a GcType when it is turned into a forwarding object.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>