Object needing finalization are detected.
Currently objects needing finalizations are kept track of and an internal
function is called whenever they need to be finalized. However, as we don't want
to call finalizers while holding the GC lock, we need to copy these objects to
some other place and run finalizers at a later time.

Additionally, we need to make sure that all finalizers are executed during
shutdown. This can quite easily be solved by just iterating through all objects
during shutdown (ignoring blocks without the finalizer flag).

Finally, there may be some cases where the finalizer flag is set even when it
doesn't need to be. This does not impact the correctness of the collector, but
may impact performance if it happens too frequently. For example, the last
generation contains a block marked as containing finalizers after running the
GcTest, even though all objects with finalizers are seemingly finalized
already. This is worth investigating!

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>