Some debugging on Linux.
It looks like we cannot use setjmp for good stack traces on Linux, as the jmpbuf does not
necessarily store everything there. For example, rbp is not stored, but is expected to be generated
from debug information during unwinding instead. This means that we need to provide our own solution
for this to make it reliable.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>