Properly update any weak references that were pinned during a collection.
Strangely, it seems that the additional thread introduced stores references to
objects on the other thread's stack, which causes excessive pinning. This causes
performance to drop and false positive rate on object watches to be fairly
high. It is possible we can not do anything about this, but it is worth
investigating a bit further.

From the stack traces it seems like the references come from within the UThread
system when calling 'sleep', but that seems unlikely and it is most likely other
stray values on the stack from the thread's main function.

We should also consider tuning the GC a bit more, as it is fairly expensive to
pause threads. Therefore, it might be worth waiting a bit more before performing
a garbage collection cycle.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>