Researched why stuttering in presentations occurred on Windows.
This seems to be due to an exception being thrown inside DirectDraw when a font
is drawn when the height is scaled to near zero. This is similar to what
happened on Linux, where Pango seemed to enter an infinite loop when drawing
text with essentially zero height. Windows did not hang at least, but throwing an
internal exception causes noticeable hiccups from time to time.

This is resolved in the presentation library by never scaling to less than 0.05
during animations, which should be sufficient for most cases.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>