Optimized handling of axis labels in presentations.
We now assume we're using a solid color, which allows us to set the transparency
directly on the brush when drawing rather than having to use a "pop" operation.

This has significant impact when running the presentation in full screen on a
Surface for example. It is now very smooth.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>