Finalizers are executed from a queue associated with each thread.
Previously, a new UThread was created for every finalizer that could not be
executed on the current thread. In some cases the GC would destroy a large
number of such objects at the same time, causing a large number of UThreads to
be spawned, eventually causing Storm to run out of virtual memory. This fix
elliminates that issue, instead starting a single UThread for all finalizers in
such a burst.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>