Fixed an underscanning issue when using SMM.
Pointers to GcType objects were not marked as GC pointers previously, as that
was not necessary when using the MPS. This is necessary with SMM, and as such
the preprocessor now marks GcType pointers as GC pointers automatically. This
required some re-writing of the special cases relating to type instances, but
they are fairly minor, and the test suite still works properly when using the
MPS, indicating that the implementation is fairly correct.

Now, Storm boots with SMM in the test suite. There are still bugs that cause it
to crash, but progress is being made.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>