Made sure to execute finalizers when shutting down SMM using 'destroy'.
It seems like *most* finalizers are executed for now. However, it is possible
that some are forgotten. When counting the number of finalizations and
finalization requests, the number consistently differs by one. This warrants
further investigation.

Still, there seems to be an error in the GcWatch implementation, as the pointer
based Set/Map tests fail sporadically when running on SMM.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>