Fixed a circular dependency in class loading that could cause issues.
Since the function "updateHandle" called "typeDesc", the type description could sometimes be created
too early, causing incorrect type descriptions and sometimes issues when a type was in the process
of being loaded.

Now, we use the "pure" flag on the functions to determine if we shall add them to the handle rather
than examining if the class is a primitive type or not. This is more accurate and does not cause the
circular dependency.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>