Fixed a consistency bug in the generation summaries in SMM.
When creating pointer summaries, the pointers were recorded before they were
handed to the actual scanning code. This caused issues when the actual scanner
updated the pointers, making the summaries contain stale information. This was
not always apparent, as the pointer summaries are usually very rough, but could
cause strange crashes from time to time. Now, the summary generation takes care
to create the summaries after the scanner have executed, so that any
modifications are taken into account properly.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>