Updated the scanner to support only scanning the header.
This means that we can scan the header of weak objects as a regular reference
even though the remainder of the object contains weak references.

In Storm, we don't rely on this behavior, as there will always be a strong
reference to all type descriptions. It is, however, nice for correctness and to
avoid much debugging work in the future if Storm ever decides to rely on weak
arrays to keep their type description alive.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>