Fixed a bug that caused inheriting from Type to fail.
THe issue was that the size was not explicitly set for core:lang:Type, which caused the size to be
computed incorrectly, and followingly the GcType generated for any types inheriting directly from
Type in Storm to be incorrect (most notably, the stride field). This is corrected now, and seems to
work properly.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>