Added a reader for licenses.
Thus, it is now possible to create .license-files without having to resort to C++ libraries. The
format is similar to C++:

<name of the license>
<names of the author(s)>
<license body>

I.e. it is line-based. The first line is a summary of the license (e.g. BSD 3-clause), the second
line is the name of the authors or organization and the remaining lines are the license body.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>