Added the C++ code for a generic database interface.
Code by: David Ångström, Markus Larsson, Mohammed Hamade, Oscar Falk, Robin
Gustavsson, Saima Akhter

Modified slightly to better adhere with the style of the rest of Storm, and to
compile on older versions of Visual Studio (pre C++11), as the rest of Storm.

More code, containing a language extension to Basic Storm, is available and will
be added shortly.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>