Making sure to preserve registers on X86 when catching exceptions.
Otherwise, not all registers were always preserved, which sometimes caused very
weird behavior in the caller of generated code which caught an exception. The
SEH mechanism does not restore any registers, so the code generation needs to be
aware that this might happen and explicitly preserve all non-clobbered registers
in the function prolog.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>