Made sure that "Type x = expr;" respects type casting rules.
Previously, "Type x = expr;" was fine to do as long as a constructor existed,
while "x = expr;" required that the constructor was marked as "autoCast". Now
both cases require auto-cast constructors to avoid potential issues.

One example of this was:

VarInfo x = result.location(s);

Where "location" previously returned "VarInfo" but now returns "Var". Since
"VarInfo" has a constructor taking a single "Var" (not a cast-ctor), this was
previously not caught by the compiler, but it is now.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>