Make sure not to destroy things that are references from other places.
In the previous model, the Variant used to capture copies of stack variables would destroy the
contents of the variable. This was good for cases when we copied the data into the variant, but not
for cases when we simply stored a reference there, as is the case with the C++ frontend.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>