Don't complain about variables that are not activated in the backend.
It is perfectly valid to have variables that are never activated. For example:

void foo() {
   return;

   Value x;
}

Here, the variable declaration is really dead code, and as such the initialization was
removed. Basic Storm does, however, currently still allocate the variable "x", and as such backends
would complain that it was never activated.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>