Fixed a minor bug in the threading library.
When calling Thread::current() from the main (Compiler) thread of the
application from a module other than the one calling Thread::setStackBase() the
threading framework would fail to fetch the current thread, causing an
assertion.

This is now fixed by properly storing the main thread in the thread local
variable as well, just as for all other threads. This should also produce fewer
branch mispredicts, as all threads are now handled the same.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>