Moved pointer comparisons to a separate package.
These needed to be patched by the new logic in Progvis. But as they were
compiled previously, Progvis had to re-load the entire package to get their
source, which took a long time as that involved parsing the entire C++ frontend
once more. Now, the relevant functions are moved into a separate package which
speeds things up significantly.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>