Using an extra NameSet during package load, so that we may revert.
Now, if an error occurs during loading, the loading progress of the packages are
reverted to the previous state by using a temporary NameSet as a "shadow space"
for loaded entities during the loaded process. This likely has a minor
performance impact (multiple hash table lookups), but should be negligible since
one of them should be mostly empty.

The same mechanism will be used for reloading packages.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>