Make sure removal in sets and maps work properly after replacing pointers.
Maps might still behave nonsensically after this, as we don't know which one out
of a set of duplicates is returned when we look for it. We could check for it on
lookup, but this could impact performance. It does not seem to be an issue
currently, as we don't have maps with Named as a key in many situations.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>