Split references to offsets from references to memory.
This was to not confuse the GC whenever offsets happen to become large (we don't
scan offsets).

The implementation seems to work satisfactory on X86, the X64 implementation has
not been tested yet. It is likely that it does not support some of the cases
where an offset should be emitted (there is a big TODO comment about this in Code/X64/Asm.cpp).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>