Using the vtable map rather than the replace map in ReplaceContext.
The bug with running heavy compilation jobs after the replacement test seem to
be due to an underscanning issue with the vtables. Probably, a vtable of an old
object is being replaced, the MPS fails to notice this (since we're abusing the
API of mps_formatted_objects_walk) and thus fails to update the reference when
the newly allocated vtable is scanned.

This is yet to be verified.

A possible workaround is to use mps_formatted_objects_walk to find all
interesting objects, add them to a list and traverse them outside the
formated_objects_walk call. Since the arena is parked at this time, we don't
have to worry about keeping all pointers in roots and the likes (i.e. we can use
an std::vector<void *>).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>