Refactored the GL backends into a uniform system relying more on Cairo.
This is since earlier testing arrived at the conclusion that relying on Cairo
for copying the rendered graphics to the final window was more reliable over
attempting to interfere with Gtk and render a window directly ourselves.

This change also allows us to use whatever Cairo backend that Gtk uses. It seems
like the OpenGL backend in Cairo causes some newer iris drivers to crash for
Intel GPUs. This change allows us to switch between software rendering, whatever
Gtk+ uses (usually hardware assisted to some degree, often using XRender) and
OpenGL.

The downside of this change is that we lose the option of skipping Gtk+ entirely
when rendering (which was not reliable anyway).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>