More tinkering with the GLL parser.
Deduplication of states as described in the paper should be done. One source of
problems at the moment is that the number of states explodes quite fast for
slightly ambiguous cases. This could be solved with a recursive "priority"
function, as in the GLR parser. If deduplication solves this anyway, this is the
preferred solution as it will reduce execution time as well!

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>