WIP: Fixed most of the compilation errors in the new GLL parser.
The last one, however, needs yet some re-structuring of the stack data.

We need to keep the match at the location where new states from rules are
spawned in order to pick the best match. We can still keep the StackFirst as we
have it, and use it to eliminate the need for the "second" parameter to
"pqPushFirst", and thereby reduce the size of the "created" array by half.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>