Started rewriting rendering in the GUI library to be single-threaded.
The Win32 implementation seems to work fine at the moment. Resizing the window
during an ongoing animation is a bit janky at the moment. The largest issue is
that animations pause when the window is moved.

The implementation on Linux is probably not working at the moment, as no effort
has been made there yet. We will have more options there, however. Hopefully, we
will be able to render directly to a window context from OpenGL, improving
performance even more in the end.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>