Fixed the "plain" text rendering function in both Skia and Cairo.
It now allocates a Text object behind the scenes. This could in theory be
allocated on the stack, but that requires additional support from the runtime to
work. This kind of short-lived objects is exactly what GCs are good at, so it
should not be a major headache, especially since the text shaping will proabbaly
be the bottleneck anyway.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>