Added licenses for Pango and Skia in preparation for release.
Also made Skia the default graphics backend on Linux. It outperforms the Cairo GL
backend by quite a lot, and the text rendering works well now.

In the future, we might want to consider dropping support for the Cairo GL
backend. That means we can skip supplying our own copy of Cairo with the GUI
library.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>