Added a "join" function usable in Storm.
Also modified how templates are handled slightly. They are now checked against
the desired logic in SimpleName as all other names are. This means that
auto-deduction of parameters in templates are allowed to happen. The template
function might simply return a "best guess" based on the available information,
and the SimpleName may then decide if it is suitable or not.

Also moved a test from the old location to the new one. Seems to have been left
after a merge.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>