Fixed a bug in the MPS integration.
If an exception occurred in a Type constructor, the GC would eventually attempt
to finalize the now destroyed object. This caused an exception to occur since
the vtable was cleared.

This should probably be a bit closer investigated to ensure that this always
happens. Otherwise, we might execute destructors more than once in some cases.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>