Fixed a bug in Basic Storm.
A simple function in a value-type object like this:

value X {
  X fn() { this; }
}

Would cause incorrect code to be generated. The implementation was confused
that "this" was a reference variable. Now it is handled both in MemberVarAccess
and in the Basic Storm Function class (i.e. fnRetRef is used if possible).

This also revealed some bugs in the fnRetRef pseudo-opcode.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>