Fixed a bug with stack traces that could lead to crashes.
When exceptions were thrown from shared libraries, multiple instances of the
StackInfoSet were created, and since all of them were not properly initialized
with the SehInfo (or DwarfInfo). This would lead to incorrect (= non-base
pointers) pointers being stored in stack traces, which in turn crashed the MPS
in some cases.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>