Re-worked some details of how the serialization works.
The largest change is that the type-info for the serialization framework now
refers to all types using a Type * value rather than a SerializedType *
value. This makes the system easier to work with from C++ as it is not necessary
to call the 'serializedType' function (which may be automatically generated in
Storm). This has another benefit. In cases where the 'serializedType' function
is implemented in C++, and it is difficult to achieve the same "caching" as in
the generated functions we would previously generate the type-info each time we
would write an instance of such a type. The new scheme allows the ObjOStream to
cache these instances, so they are only created once per stream.

Additionally, a new header "Core/Io/SerializationUtils.h" was added. It contains
some C++ templates that makes it convenient to use the serialization library in
C++ as well (Nat::write is not valid in C++ for example). Now, we can do
Serialize<Nat>::read() instead, which even works for arrays and other types (in
the future) that have their serialization code generated at runtime.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>