The "async" keyword now works like regular function calls wrt cloning.
Previously, calling a function with the "async" keyword would always cause any
non-actors to be copied. This was inconsistent compared to how regular function
calls worked, where parameters would only be copied if the compiler is unable to
prove that both were to execute on the same thread.

Now, "spawn" behaves in the same way. Both with respect to parameters and return
values. As a consequence of this change, the interface to the Function class has
been refactored to work similarly for regular function calls and async
calls. The Future object has also gained the ability to not clone its results.

Note that values are still copied more times when doing an async call compared
to a regular function call. This is, however, unavoidable and not a
problem (copying a value is supposed to be cheap and possible to do on a whim).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>