Fixed a bug in os::Sema.
When a os::Sema was used to guard wakes inside the repaint logic in Progvis, the
semaphore could be deallocated before os::Sema::up() would have time to release
the lock. This only seemed to happen on single-processor systems.

Now the os::Sema makes sure to wake any threads as the last thing it does, so
this problem is avoided. Using a semaphore in this way (i.e. deallocating it
soon after os::Sema::down() has returned) is fairly common.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>