Allow Basic Storm to differentiate between proper "this" and fake "this".
Since it is possible to declare custom variables with the name "this", Basic
Storm needs to be able to differentiate between such variables created by the
system and by the user. Variables created by the system are constant, and Basic
Storm makes assumption about them. In particular, generated "this" variables are
known to be callable without a thread switch always. This is not true for
user-defined "this" variables (which are just a syntactic convenience for
non-member functions).

This introduces a special value for ValParam in Basic Storm that is created by
the helper function "thisParam()". This distinction needs to be made in all
code that uses Basic Storm to generate code. In general it only matters when
Actors that are not statically bound to a thread is used. This commit updates
all occurrences in Basic Storm code in this source tree.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>