Rebuilt how continuous updates are done on Gtk.
Now, we are using the tick callback in Gtk to drive continuous updates in sync
with the frame clock. This minimizes the inter-thread communication required to
post updates to the UI thread, and thus poses less risk of "losing" an update
due to a repaint request arriving at the "wrong" time. We still need updates
from the first repaint though, and we could probably revise the implementation
further to make it even more robust.

Signed-off-by: Filip Strömbäck <filip.stromback@liu.se>