Exports are working for syntax.
Having different export "types" like this might lead to confusion, so we might
re-think the export system at a future point. If we only have a single type of
exports, we can also make the Package handle exports. This means that other
languages would not need to care to pay attention to exports. In this case, we
should have a separate file with exports (named "exports", perhaps) that just
lists exported packages in order to avoid circular dependencies (exports are
needed for figuring out the syntax to use for the current file, for example).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>