Fix issue where pointers to threads stacks don't show properly.
This means that data structures "below" a thread will have their pointers shown
on top of the stack, even though none of them refer to the thread. I think this
is a suitable tradeoff, as properly handling this would require the ability to
treat pointers to a thread's stack separately.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>