Create templates more eagerly.
Previously, a template would not be created if another named entity would match,
even though a template could provide a better match. This would even happen if
the original matching would result in two or more named entities with the same
badness.

Now, we try to create templates more eagerly, always if we find a match that is
not considered perfect. This solves the above issue, and makes templates act
more as if they are always present since we are better at hiding the fact that
we create them on the fly.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>