Revised the back-edge trakcing.
Now it is based on the already existing "location" callback. Thus, we will have
an experience closer to what the user sees, rather than how it is
implemented (although it is not a perfect match, it is better than the previous
approach). As a bonus, it has a lot less overhead.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>