Fixed a bug in Core/Future.
The underlying future object was allocated in a GC-object, but the pointer was
not properly exposed to the GC, which could therefore cause crashes.

The implementation is now rewritten so that we both get the safety of having all
memory allocated by the GC (i.e. no memory leaks), but still the more prompt
release of resources from refcounting the static allocation (the static
allocation is typically in a pool that is collected less frequently than the
pool where the FutureBase is allocated).

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>