Fixed the bug in the model checker for Progvis.
The bug was that the wait queue for locks, semaphores and condition variables
was not properly exposed to the checker. As such, it would incorrectly merge two
states where a different thread was next in line to wake. We now expose this
information to the checker so that it works properly.

This also revealed that we currently need to number threads by their thread-id,
and not their sequence number in the summary. Otherwise, when a thread
terminates we might try to step the wrong thread-id. This is possible to solve.

Signed-off-by: Filip Strömbäck <fstromback@gmail.com>