Make sure that a barrier step in Progvis does not terminate a thread.
When a thread executes a function without any barriers, a barrier step would
previously cause the thread to exit immediately, meaning that any memory
accesses made by the thread would not be visible. Now, it stops just before
exiting instead, so that the memory accesses are visible. This means that the
auto-grader can find such issues as well.